feat(domain): add AggregateRoot
An AGGREGATE is a cluster of
associated objects that we treat as a unit for the purpose of data
changes. Each AGGREGATE has a root and a boundary. The boundary defines
what is inside the AGGREGATE. The root is a single, specific ENTITY
contained in the AGGREGATE.

The root is the only member of the
AGGREGATE that outside objects are allowed to hold references to[.]

- support to record a domain event
- support to pull domain events

BREAKING CHANGE: AggregateRoot Class was added